requirements.txt
# ==============================
# Web server
# ==============================
Django==1.11.26

# Consider updating these for python 3 support:
gunicorn==19.5
dj-static==0.0.6


# ==============================
# Background processing
# ==============================

celery==3.1.23
django-celery==3.1.17



# ==============================
# Database and Persistence
# ==============================

# DB Settings and Management
psycopg2==2.8.4
psycogreen==1.0
# consider updates to 0.5.0 for python 3 support:
dj-database-url==0.4.1
django-object-actions==1.0.0

# Caching
django-redis==4.4.2
django-cache-url==1.0.0

# File Storage
boto==2.39.0
django-storages==1.6.6

# hierarchy support for our dataset labels:
# TODO: move over to django-mptt in Python 3.5+:
# https://github.com/django-mptt/django-mptt
django-closuretree==1.2.0


# ==============================
# User Accounts and Social Media
# ==============================

# TODO: Using temporary fork until PR is merged:
# https://github.com/python-social-auth/social-core/pull/384
-e git+https://github.com/modulitos/social-core.git@31460b799752630d93e872f5bb50225f7e590a80#egg=social_auth_core
-e git+https://github.com/modulitos/social-app-django.git@99600f11090b874b190106efe9901b560d275b98#egg=social_auth_app_django

django-loginas==0.2.2
django-oauth-toolkit== 1.0.0


django-cors-headers==2.5.2
# last version with Django 1.10 support:



# ==============================
# REST API
# ==============================

djangorestframework==3.7.7
djangorestframework-csv==2.0.0

# git+https://github.com/mjumbewu/django-rest-framework-bulk.git@84a5d6c#egg=djangorestframework-bulk==0.1.3
# replace forked drf-bulk with main version
djangorestframework-bulk==0.2
six>=1.10.0
markdown==2.6.11  # For browsable API docs
python-dateutil==2.5
ujson==1.35
bleach==1.4.3

# The Django admin interface
# TODO: upgrade to 1.0.4 for django 1.11:
django-ace==1.0.4
django-nested-admin==3.2.3
django-admin-sortable2==0.7

# The manager interface
requests>=2.9.1

# Support for JSON Web Token generation
pyjwt==1.7.1



# ==============================
# Testing, validating, and debugging
# ==============================

nose==1.3.7
django-nose==1.4.6
mock==1.3.0
# delete this? Unused?
responses==0.5.1
# delete this? Unused?
django-debug-toolbar==1.4
# TODO: upgrade to 6.x for django 1.11:
raven==6.10.0  # For Sentry error logging

# code formatter
black==19.10b0


# - - - - - - - - - - - - - - - -

# For DRF 0.4 (deprecated)
URLObject>=2.4.0

